Move helpers to separarate classes
